requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.6
exceptiongroup==1.1.2
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
pycparser==2.21
PyJWT==2.7.0
pytest==7.4.0
pytest-django==4.5.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
